ml_pipeline/cloud_functions/trigger_skew/requirements.txt
functions-framework==3.*
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
